PIQA-68 Purpose Consumer Listing (#85)
