+++ features/purpose/step_definitions/purpose-listing-producer.ts
@@ -1,18 +1,13 @@
-import assert from "assert";
-import { Given, Then, When } from "@cucumber/cucumber";
+import { When } from "@cucumber/cucumber";
 import { z } from "zod";
 import {
   assertContextSchema,
   getAuthorizationHeader,
   getOrganizationId,
-  getRandomInt,
-  getRiskAnalysis,
-  getToken,
 } from "../../../utils/commons";
 import { apiClient } from "../../../api";
 import { TenantType } from "../../common-steps";
 import { PurposeVersionState } from "../../../api/models";
-import { dataPreparationService } from "../../../services/data-preparation.service";
 
 When(
   "l'utente erogatore richiede una operazione di listing delle finalità limitata alle prime {int} finalità",
@@ -27,22 +22,6 @@ When(
       },
       getAuthorizationHeader(this.token)
     );
-  }
-);
-
-Then(
-  "si ottiene status code {int} e la lista di {int} finalità",
-  async function (statusCode: number, count: number) {
-    assertContextSchema(this, {
-      response: z.object({
-        status: z.number(),
-        data: z.object({
-          results: z.array(z.unknown()),
-        }),
-      }),
-    });
-    assert.equal(this.response.status, statusCode);
-    assert.equal(this.response.data.results.length, count);
   }
 );
 
@@ -151,41 +130,6 @@ When(
       },
       getAuthorizationHeader(this.token)
     );
-  }
-);
-
-Given(
-  "{string} ha già creato una finalità in stato {string} per quell'e-service contenente la keyword {string}",
-  async function (
-    tenantType: TenantType,
-    purposeState: PurposeVersionState,
-    keyword: string
-  ) {
-    assertContextSchema(this, {
-      eserviceId: z.string(),
-    });
-
-    const token = await getToken(tenantType);
-    const consumerId = getOrganizationId(tenantType);
-
-    const { riskAnalysisForm } = getRiskAnalysis({
-      completed: true,
-      tenantType,
-    });
-
-    const title = `purpose ${this.TEST_SEED} - ${getRandomInt()} - ${keyword}`;
-    this.purposeId = await dataPreparationService.createPurposeWithGivenState({
-      token,
-      testSeed: this.TEST_SEED,
-      eserviceMode: "DELIVER",
-      payload: {
-        title,
-        eserviceId: this.eserviceId,
-        consumerId,
-        riskAnalysisForm,
-      },
-      purposeState,
-    });
   }
 );
